# spec/openapi.yaml
openapi: 3.1.0
info:
  title: Гармония (Harmony) API
  version: 1.0.0
  summary: REST and realtime API for the Harmony social network.
  description: |
    # Гармония API

    A social network API. All responses share one envelope so a client has a
    single code path for success and failure:

    ```json
    { "ok": true,  "data": { }, "meta": { "count": 15, "has_more": true, "next_cursor": "eyJ..." } }
    { "ok": false, "error": { "code": "not_found", "message": "Пост не найден.", "request_id": "..." } }
    ```

    ## Authentication

    After `POST /auth/login` the API issues two cookies:

    | Cookie | HttpOnly | Purpose |
    |---|---|---|
    | `harmony_access` | yes | Access token (JWT) |
    | `harmony_refresh` | yes | Refresh token, scoped to `/api/v1/auth` |
    | `harmony_csrf` | **no** | CSRF token for the double-submit pattern |

    The access cookie is `HttpOnly` so that an XSS payload cannot read it.
    The CSRF cookie is deliberately readable, because the client must echo it.

    **Every unsafe method** (`POST`, `PATCH`, `PUT`, `DELETE`) must send the CSRF
    token in the `X-CSRF-Token` header, matching the cookie. A correct `Origin`
    or `Referer` is also accepted, which keeps non-browser clients working.

    Non-browser clients may instead send `Authorization: Bearer <access_token>`.

    ## Resources

    Resources are addressed by an opaque 32-character `public_id`, never by the
...
    bugs easier to exploit.

    ## Pagination

...

...

...

...

...

...

...

  contact:
...
servers:
...
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /posts/{post_id}/comments:
    get:
      tags: [Comments]
      summary: Comments on a post
      security: []
      parameters:
        - $ref: "#/components/parameters/PostId"
        - name: sort
          in: query
          schema: { type: string, enum: [top, newest], default: top }
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Comments, each with a short preview of its replies
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Comment" }
                  meta: { $ref: "#/components/schemas/PageMeta" }
        "404": { description: Post not found }
    post:
      tags: [Comments]
      summary: Comment on a post
      parameters:
        - $ref: "#/components/parameters/PostId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body: { type: string, maxLength: 2000 }
                parent_id: { type: string, description: Public id of the comment being replied to. }
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: object
                    properties:
                      comment: { $ref: "#/components/schemas/Comment" }
                    additionalProperties: true
        "422":
          description: "Empty, or reply depth exceeded"
        "451": { description: "Rejected by content policy" }
  /comments/{comment_id}:
    patch:
      tags: [Comments]
      summary: Edit a comment
      parameters:
        - $ref: "#/components/parameters/CommentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties: { body: { type: string, maxLength: 2000 } }
      responses:
        "200": { description: Updated }
        "403": { description: Not the author }
    delete:
      tags: [Comments]
      summary: Delete a comment
      parameters:
        - $ref: "#/components/parameters/CommentId"
      responses:
        "204": { description: Deleted }
        "403": { description: Not the author }
  /comments/{comment_id}/replies:
    get:
      tags: [Comments]
      summary: Replies to a comment
      security: []
      parameters:
        - $ref: "#/components/parameters/CommentId"
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 30 }
      responses:
        "200":
          description: Replies
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Comment" }
                  meta: { $ref: "#/components/schemas/PageMeta" }

  # ------------------------------------------------------------- Uploads ---
...
  /conversations:
    get:
      tags: [Chat]
      summary: List conversations
      responses:
        "200":
          description: Conversations
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: object
                    properties:
                      total_unread: { type: integer }
                      conversations:
                        type: array
                        items: { $ref: "#/components/schemas/Conversation" }
    post:
      tags: [Chat]
      summary: Start or open a direct conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id]
              properties:
                user_id: { type: string, description: Public id of the recipient. }
      responses:
        "201":
          description: Conversation ready
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: object
                    properties:
                      created: { type: boolean }
                      conversation: { $ref: "#/components/schemas/Conversation" }
        "403":
          description: Recipient does not accept messages from strangers
        "409": { description: Recipient unavailable }
  /conversations/{conversation_id}:
    get:
      tags: [Chat]
      summary: Read a conversation
      parameters:
        - $ref: "#/components/parameters/ConversationId"
      responses:
        "200":
          description: Conversation
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: object
                    properties:
                      conversation: { $ref: "#/components/schemas/Conversation" }
                    additionalProperties: true
        "404": { description: "Not a participant" }
  /conversations/{conversation_id}/messages:
    get:
      tags: [Chat]
      summary: Message history
      description: Oldest first, so the result renders directly.
      parameters:
        - $ref: "#/components/parameters/ConversationId"
        - $ref: "#/components/parameters/Cursor"
        - name: limit
          in: query
          schema: { type: integer, default: 40, maximum: 100 }
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Message" }
                  meta: { $ref: "#/components/schemas/PageMeta" }
    post:
      tags: [Chat]
      summary: Send a message (REST fallback)
      description: The WebSocket `message:send` event is preferred; this endpoint exists for clients that cannot hold a socket open.
      parameters:
        - $ref: "#/components/parameters/ConversationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties: { body: { type: string, maxLength: 4000 } }
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: object
                    properties:
                      message: { $ref: "#/components/schemas/Message" }
                    additionalProperties: true
        "429": { description: "Message rate limit" }
        "451": { description: "Rejected by content policy" }
  /conversations/{conversation_id}/read:
    post:
      tags: [Chat]
      summary: Mark as read
      parameters:
        - $ref: "#/components/parameters/ConversationId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                up_to_message_id: { type: string }
      responses:
        "200": { description: Marked }
  /conversations/{conversation_id}/leave:
    post:
      tags: [Chat]
      summary: Leave a conversation
      parameters:
        - $ref: "#/components/parameters/ConversationId"
      responses:
        "204": { description: Left }
  /messages/{message_id}:
    delete:
      tags: [Chat]
      summary: Delete a message
      parameters:
        - name: messageId
          in: path
          required: true
          schema: { type: string }
      responses:
        "204": { description: Deleted }
        "403": { description: Not the sender }
...
  /notifications:
    get:
      tags: [Notifications]
      summary: Inbox
      parameters:
        - name: unread
          in: query
          schema: { type: string, enum: ["1"] }
        - $ref: "#/components/parameters/Cursor"
        - name: limit
          in: query
          schema: { type: integer, default: 30 }
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Notification" }
                  meta: { $ref: "#/components/schemas/PageMeta" }
...
  /notifications/read:
    post:
      tags: [Notifications]
      summary: Mark as read
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                ids:
                  type: array
                  items: { type: string }
                all: { type: boolean }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: object
                    properties:
                      marked: { type: integer }
                      unread_count: { type: integer }
  /notifications/{notification_id}:
    delete:
      tags: [Notifications]
      summary: Delete a notification
      parameters:
        - name: notificationId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: Deleted }
        "404": { description: Not found }

  # ---------------------------------------------------------- Moderation ---
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: harmony_access
      description: |
        Access token set by `/auth/login`. Send `X-CSRF-Token` on unsafe methods.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: For non-browser clients. Exempt from CSRF.
  parameters:
...
    PostId:
      name: postId
      in: path
      required: true
      schema: { type: string, pattern: '^[0-9a-f]{32}$' }
    CommentId:
      name: commentId
      in: path
      required: true
      schema: { type: string, pattern: '^[0-9a-f]{32}$' }
    ConversationId:
      name: conversationId
      in: path
      required: true
      schema: { type: string, pattern: '^[0-9a-f]{32}$' }
...
    Cursor:
      name: cursor
      in: query
      description: Opaque keyset cursor from `meta.next_cursor`.
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, default: 15, minimum: 1, maximum: 100 }
  schemas:
...
    PageMeta:
      type: object
      properties:
        count: { type: integer }
        has_more: { type: boolean }
        next_cursor: { type: [string, "null"] }
        total: { type: integer }
...
    User:
      type: object
      description: |
        The public projection. Never includes the email unless the target has
        opted in, nor any moderation state, counter or token.
      properties:
        public_id: { type: string }
        username: { type: string }
        display_name: { type: [string, "null"] }
        initials: { type: string }
        bio: { type: [string, "null"] }
        avatar_url: { type: [string, "null"] }
        avatar_color: { type: string }
        location: { type: [string, "null"] }
        website: { type: [string, "null"] }
        pronouns: { type: [string, "null"] }
        role: { type: string, enum: [member, moderator, admin] }
        is_verified: { type: boolean, description: Active account with a confirmed email. }
        posts_count: { type: integer }
        comments_count: { type: integer }
        followers_count: { type: integer }
        following_count: { type: integer }
        created_at: { type: string, format: date-time }
        last_seen_at: { type: string, format: date-time }
        email: { type: string, description: Present only if the target enables it. }
        profile_visibility: { type: string }
        is_self: { type: boolean }
...
    Comment:
      type: object
      properties:
        id: { type: string }
        post_id: { type: string }
        parent_id: { type: [string, "null"] }
        depth: { type: integer, maximum: 3 }
        author: { $ref: "#/components/schemas/User" }
        body: { type: string }
        status: { type: string }
        likes_count: { type: integer }
        replies_count: { type: integer }
        replies:
          type: array
          items: { $ref: "#/components/schemas/Comment" }
        is_owner: { type: boolean }
        created_at: { type: string, format: date-time }
        edited_at: { type: [string, "null"], format: date-time }
    Conversation:
      type: object
      properties:
        id: { type: string }
        kind: { type: string, enum: [direct, group] }
        title: { type: [string, "null"] }
        participants:
          type: array
          items: { $ref: "#/components/schemas/User" }
        messages_count: { type: integer }
        last_message_at: { type: [string, "null"], format: date-time }
        last_message_preview: { type: [string, "null"] }
        unread_count: { type: integer }
        last_read_at: { type: [string, "null"], format: date-time }
        is_muted: { type: boolean }
        created_at: { type: string, format: date-time }
    Message:
      type: object
      properties:
        id: { type: string }
        conversation_id: { type: string }
        sender: { $ref: "#/components/schemas/User" }
        reply_to_id: { type: [string, "null"] }
        body: { type: string }
        status: { type: string, enum: [queued, sent, blocked, deleted] }
        is_deleted: { type: boolean }
        created_at: { type: string, format: date-time }
        edited_at: { type: [string, "null"], format: date-time }
    Notification:
      type: object
      properties:
        id: { type: string }
        kind: { type: string, enum: [like, comment, reply, follow, message, moderation, warning, system, reminder] }
        title: { type: string }
        body: { type: string }
        url: { type: [string, "null"] }
        resource_type: { type: [string, "null"] }
        resource_id: { type: [string, "null"] }
        is_read: { type: boolean }
        actor: { $ref: "#/components/schemas/User" }
        created_at: { type: string, format: date-time }
...